requirements.txt
# InstaSplitter dependencies
Pillow>=10.0.0
